links DSL: Some changes resulting from using it

+++ hashdist/core/links.py
@@ -7,7 +7,7 @@ uses it to create (a potentially large number of) links.
 
 The following rules creates links to everything in "/usr/bin", except
 for "/usr/bin/gcc-4.6" which is copied (though it could be achieved
-more easily with the `force` flag)::
+more easily with the `overwrite` flag)::
 
   [
     {
@@ -28,11 +28,19 @@ more easily with the `force` flag)::
     
   ]
 
-Rules are applied in order.
+Rules are executed in order. If a target file already exists, nothing
+happens.
+
 
 **action**:
-  One of "symlink", "copy", "exclude", "launcher". Other types may be added
-  later.
+  One of "absolute_symlink", "relative_symlink", "copy", "exclude",
+  "launcher". Other types may be added later.
+
+  * *absolute_symlink* creates absolute symlinks
+  * *relative_symlink* creates relative symlinks
+  * *copy* copies contents and mode (``shutil.copy``)
+  * *exclude* makes sure matching files are not considered in rules below
+  * *launcher*, see :func:`make_launcher`
 
 **select**, **prefix**:
   `select` contains glob of files to
@@ -54,8 +62,13 @@ Rules are applied in order.
   Target filename (`source` is used) or directory (`select` is used).
   Variable substitution is performed.
 
-**force**:
+**dirs**:
+  If present and `True`, symlink matching directories, not only files.
+  Only takes effect for `select`; `source` always selectes dirs.
+
+**overwrite**:
   If present and `True`, overwrite target.
+
 
 """
 
@@ -66,34 +79,70 @@ import shutil
 import errno
 from string import Template
 
-from .fileutils import silent_makedirs, silent_unlink
+from .fileutils import (silent_makedirs, silent_unlink, silent_relative_symlink,
+                        silent_absolute_symlink, silent_copy)
 from ..hdist_logging import null_logger
 
-from .ant_glob import glob_files
+from .ant_glob import ant_iglob
 
 
 def expandtemplate(s, env):
     return Template(s).substitute(env)
 
 def make_launcher(src, dst, launcher_program):
+    """
+    The 'launcher' action. This is a general tool for processing
+    bin-style directories where the binaries must have a "fake"
+    argv[0] target (read: Python). The action depends on the source type:
+
+    program (i.e., executable not starting with #!):
+        Set up as symlink to "launcher", which is copied into same directory;
+        and "$dst.link" is set up to point relatively to "$src".
+
+    symlink:
+        Copy it verbatim. Thus, e.g., ``python -> python2.7'' will point to the
+        newly, "launched-ified" ``python2.7''.
+
+    other (incl. scripts):
+        Symlink relatively to it.
+    """
     dstdir = os.path.dirname(dst)
-    dst_launcher = pjoin(dstdir, 'launcher')
-    if not os.path.exists(dst_launcher):
-        shutil.copyfile(launcher_program, dst_launcher)
-    with open(dst + '.link', 'w') as f:
-        f.write(os.path.relpath(src, dstdir))
-    os.symlink('launcher', dst)
-    
+
+    type = 'other'
+    if os.path.islink(src):
+        type = 'symlink'
+    elif os.stat(src).st_mode & 0o111:
+        with open(src) as f:
+            if f.read(2) != '#!':
+                type = 'program'
+
+    if type in 'symlink':
+        os.symlink(os.readlink(src), dst)
+    elif type == 'program':
+        if launcher_program is None or not os.path.exists(launcher_program):
+            raise TypeError('Did not provide path to "launcher" program')
+        dst_launcher = pjoin(dstdir, 'launcher')
+        if not os.path.exists(dst_launcher):
+            shutil.copy(launcher_program, dst_launcher)
+        with open(dst + '.link', 'w') as f:
+            f.write(os.path.relpath(src, dstdir))
+        os.symlink('launcher', dst)
+    else:
+        os.symlink(os.path.relpath(src, dstdir), dst)
         
 
-_ACTIONS = {'symlink': os.symlink, 'copy': shutil.copyfile, 'launcher': make_launcher}
+_ACTIONS = {'symlink': silent_absolute_symlink,
+            'relative_symlink': silent_relative_symlink,
+            'absolute_symlink': silent_absolute_symlink,
+            'copy': silent_copy,
+            'launcher': make_launcher}
 
-def _put_actions(makedirs_cache, action_name, force, source, dest, actions):
+def _put_actions(makedirs_cache, action_name, overwrite, source, dest, actions):
     path, basename = os.path.split(dest)
     if path != '' and path not in makedirs_cache:
         actions.append((silent_makedirs, path))
         makedirs_cache.add(path)
-    if force:
+    if overwrite:
         actions.append((silent_unlink, dest))
     try:
         actions.append((_ACTIONS[action_name], source, dest))
@@ -108,7 +157,7 @@ def _glob_actions(rule, excluded, makedirs_cache, env, actions):
     selected = set()
     for pattern in select:
         pattern = expandtemplate(pattern, env)
-        selected.update(glob_files(pattern, ''))
+        selected.update(ant_iglob(pattern, '', include_dirs=rule.get('dirs', False)))
     selected.difference_update(excluded)
     if len(selected) == 0:
         return
@@ -131,7 +180,7 @@ def _glob_actions(rule, excluded, makedirs_cache, env, actions):
                 raise ValueError('%s does not start with %s' % (p, prefix))
             remainder = p[len(prefix):]
             target = pjoin(target_prefix, remainder)
-            _put_actions(makedirs_cache, action_name, rule.get('force', False),
+            _put_actions(makedirs_cache, action_name, rule.get('overwrite', False),
                          p, target, actions)
 
 def _single_action(rule, excluded, makedirs_cache, env, actions):
@@ -142,7 +191,7 @@ def _single_action(rule, excluded, makedirs_cache, env, actions):
         excluded.add(source)
     else:
         target = expandtemplate(rule['target'], env)
-        _put_actions(makedirs_cache, rule['action'], rule.get('force', False),
+        _put_actions(makedirs_cache, rule['action'], rule.get('overwrite', False),
                      source, target, actions)
 
 def dry_run_links_dsl(rules, env={}):
@@ -205,7 +254,8 @@ def execute_links_dsl(rules, env={}, launcher_program=None, logger=null_logger):
     logger : Logger
 
     """
-    for action in dry_run_links_dsl(rules, env):
+    actions = dry_run_links_dsl(rules, env)
+    for action in actions:
         action_desc = "%s%r" % (action[0].__name__, action[1:])
         try:
             if action[0] is make_launcher:
